use 64-bit off_t for file sizes

+++ src/misc_tools.c
@@ -313,12 +313,12 @@ bool file_exists(const char *path)
 }
 
 /* returns file size or -1 on error */
-uint64_t file_size(const char *path)
+off_t file_size(const char *path)
 {
     struct stat st;
 
     if (stat(path, &st) == -1)
         return -1;
 
-    return (uint64_t) st.st_size;
+    return st.st_size;
 }
